Migrate to a decorator function
