+++ python/pyspark/sql/worker/python_streaming_sink_runner.py
@@ -20,7 +20,6 @@ import sys
 from typing import IO
 
 from pyspark.accumulators import _accumulatorRegistry
-from pyspark.debug import FaultHandlerIntegration
 from pyspark.errors import PySparkAssertionError
 from pyspark.logger.worker_io import capture_outputs
 from pyspark.serializers import (
@@ -35,7 +34,7 @@ from pyspark.sql.types import (
     _parse_datatype_json_string,
     StructType,
 )
-from pyspark.util import handle_worker_exception, local_connect_and_auth
+from pyspark.util import handle_worker_exception, local_connect_and_auth, with_fault_handler
 from pyspark.worker_util import (
     check_python_version,
     read_command,
@@ -48,6 +47,7 @@ from pyspark.worker_util import (
 )
 
 
+@with_fault_handler
 def main(infile: IO, outfile: IO) -> None:
     """
     Main method for committing or aborting a data source streaming write operation.
@@ -57,10 +57,7 @@ def main(infile: IO, outfile: IO) -> None:
     responsible for invoking either the `commit` or the `abort` method on a data source
     writer instance, given a list of commit messages.
     """
-    fault_handler_integration = FaultHandlerIntegration()
     try:
-        fault_handler_integration.start()
-
         check_python_version(infile)
 
         setup_spark_files(infile)
@@ -131,8 +128,6 @@ def main(infile: IO, outfile: IO) -> None:
     except BaseException as e:
         handle_worker_exception(e, outfile)
         sys.exit(-1)
-    finally:
-        fault_handler_integration.stop()
 
     send_accumulator_updates(outfile)
