+++ python/pyspark/util.py
@@ -18,6 +18,7 @@
 
 import copy
 import functools
+import faulthandler
 import itertools
 import os
 import platform
@@ -915,6 +916,56 @@ def default_api_mode() -> str:
         return "connect"
     else:
         return "classic"
+
+class _FaultHandlerIntegration:
+    def __init__(self):
+        self._log_path = None
+        self._log_file = None
+        self._periodic_dump = False
+
+    def start(self):
+        self._log_path = os.environ.get("PYTHON_FAULTHANDLER_DIR", None)
+        tracebackDumpIntervalSeconds = os.environ.get(
+            "PYTHON_TRACEBACK_DUMP_INTERVAL_SECONDS", None
+        )
+        if self._log_path:
+            self._log_path = os.path.join(self._log_path, str(os.getpid()))
+            self._log_file = open(self._log_path, "w")
+
+            faulthandler.enable(file=self._log_file)
+
+            if tracebackDumpIntervalSeconds is not None and int(tracebackDumpIntervalSeconds) > 0:
+                self._periodic_dump = True
+                faulthandler.dump_traceback_later(int(tracebackDumpIntervalSeconds), repeat=True)
+
+    def stop(self):
+        if self._periodic_dump:
+            faulthandler.cancel_dump_traceback_later()
+            self._periodic_dump = False
+        if self._log_path:
+            faulthandler.disable()
+            if self._log_file:
+                self._log_file.close()
+                self._log_file = None
+            os.remove(self._log_path)
+            self._log_path = None
+
+
+def with_fault_handler(func: Callable):
+    """
+    Registers fault handler for the duration of function execution.
+    After function execution is over the faulthandler registration is cleaned as well,
+    including any files created for the integration. Not reentrant.
+    """
+    @functools.wraps(func)
+    def wrapper(*args: Any, **kwargs: Any) -> Any:
+        fault_handler = _FaultHandlerIntegration()
+        try:
+            fault_handler.start()
+            return func(*args, **kwargs)
+        finally:
+            fault_handler.stop()
+    return wrapper
 
 
 if __name__ == "__main__":
